:art: 디렉토리 구성 및 공유
