+++ src/main/java/com/semi/lynk/common/MainController.java
@@ -1,0 +1,13 @@
+package com.semi.lynk.common;
+
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.GetMapping;
+
+@Controller
+public class MainController {
+
+    @GetMapping("/")
+    public String list (){
+        return "common/main";
+    }
+}
