--- src/main/java/com/semi/lynk/controller/MainController.java
@@ -1,13 +1,0 @@
-package com.semi.lynk.controller;
-
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
-
-@Controller
-public class MainController {
-
-    @GetMapping("list")
-    public String list (){
-        return "/";
-    }
-}
