+++ src/main/java/com/semi/lynk/function/db_management/controller/DbController.java
@@ -1,0 +1,4 @@
+package com.semi.lynk.function.db_management.controller;
+
+public class DbController {
+}
