+++ src/main/java/com/semi/lynk/function/db_management/model/dao/DbMapper.java
@@ -1,0 +1,4 @@
+package com.semi.lynk.function.db_management.model.dao;
+
+public interface DbMapper {
+}
