+++ src/main/java/com/semi/lynk/function/db_management/model/dto/DbDTO.java
@@ -1,0 +1,4 @@
+package com.semi.lynk.function.db_management.model.dto;
+
+public class DbDTO {
+}
