+++ src/main/java/com/semi/lynk/function/db_management/service/DbService.java
@@ -1,0 +1,4 @@
+package com.semi.lynk.function.db_management.service;
+
+public class DbService {
+}
