+++ src/main/java/com/semi/lynk/function/electronic_payment/controller/PaymentController.java
@@ -1,0 +1,4 @@
+package com.semi.lynk.function.electronic_payment.controller;
+
+public class PaymentController {
+}
