+++ src/main/java/com/semi/lynk/function/electronic_payment/model/dao/PaymentMapper.java
@@ -1,0 +1,4 @@
+package com.semi.lynk.function.electronic_payment.model.dao;
+
+public interface PaymentMapper {
+}
