+++ src/main/java/com/semi/lynk/function/electronic_payment/model/dto/PaymentDTO.java
@@ -1,0 +1,4 @@
+package com.semi.lynk.function.electronic_payment.model.dto;
+
+public class PaymentDTO {
+}
