+++ src/main/java/com/semi/lynk/function/electronic_payment/service/PaymentService.java
@@ -1,0 +1,4 @@
+package com.semi.lynk.function.electronic_payment.service;
+
+public class PaymentService {
+}
