+++ src/main/java/com/semi/lynk/function/human/model/dao/EmployeeMapper.java
@@ -1,0 +1,17 @@
+package com.semi.lynk.function.human.model.dao;
+
+import com.semi.lynk.function.human.model.dto.EmpAndDepDTO;
+import com.semi.lynk.function.human.model.dto.EmployeeDTO;
+import com.semi.lynk.function.human.model.dto.HumanDTO;
+import org.apache.ibatis.annotations.Mapper;
+
+import java.util.List;
+
+@Mapper
+public interface EmployeeMapper {
+    List<EmployeeDTO> employeeFullList();
+
+    List<EmpAndDepDTO> joinListResult();
+
+    int employeeRegist(HumanDTO humanDTO);
+}
