+++ src/main/java/com/semi/lynk/function/human/model/dto/DepartmentDTO.java
@@ -1,0 +1,15 @@
+package com.semi.lynk.function.human.model.dto;
+
+import lombok.*;
+
+@NoArgsConstructor
+@AllArgsConstructor
+@Setter
+@Getter
+@ToString
+public class DepartmentDTO {
+
+    private int depNo;      // 부서번호
+    private String depName; // 부서이름
+
+}
