+++ src/main/java/com/semi/lynk/function/human/service/EmployeeService.java
@@ -1,0 +1,40 @@
+package com.semi.lynk.function.human.service;
+
+import com.semi.lynk.function.human.model.dao.EmployeeMapper;
+import com.semi.lynk.function.human.model.dto.EmpAndDepDTO;
+import com.semi.lynk.function.human.model.dto.EmployeeDTO;
+import com.semi.lynk.function.human.model.dto.HumanDTO;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+import java.util.List;
+
+@Service
+public class EmployeeService {
+
+    private EmployeeMapper mapper;
+
+    @Autowired
+    public EmployeeService (EmployeeMapper mapper) {
+        this.mapper = mapper;
+    }
+
+    public List<EmployeeDTO> employeeList () {
+
+        return mapper.employeeFullList();
+    }
+
+    public List<EmpAndDepDTO> joinList() {
+
+        return mapper.joinListResult();
+    }
+
+    // DML용
+    @Transactional // 정상 처리 : commit , 예외 발생 : rollback
+    public int humanResister(HumanDTO humanDTO) {
+
+        int result = mapper.employeeRegist(humanDTO);
+        return result;
+    }
+}
