+++ src/main/java/com/semi/lynk/function/login/controller/LoginController.java
@@ -1,0 +1,4 @@
+package com.semi.lynk.function.login.controller;
+
+public class LoginController {
+}
