+++ src/main/java/com/semi/lynk/function/login/model/dao/LoginMapper.java
@@ -1,0 +1,4 @@
+package com.semi.lynk.function.login.model.dao;
+
+public interface LoginMapper {
+}
