+++ src/main/java/com/semi/lynk/function/login/model/dto/LoginDTO.java
@@ -1,0 +1,4 @@
+package com.semi.lynk.function.login.model.dto;
+
+public class LoginDTO {
+}
