+++ src/main/java/com/semi/lynk/function/login/service/LoginService.java
@@ -1,0 +1,4 @@
+package com.semi.lynk.function.login.service;
+
+public class LoginService {
+}
