+++ src/main/java/com/semi/lynk/function/notice_board/controller/NoticeController.java
@@ -1,0 +1,4 @@
+package com.semi.lynk.function.notice_board.controller;
+
+public class NoticeController {
+}
