+++ src/main/java/com/semi/lynk/function/notice_board/model/dao/NoticeMapper.java
@@ -1,0 +1,4 @@
+package com.semi.lynk.function.notice_board.model.dao;
+
+public interface NoticeMapper {
+}
