+++ src/main/java/com/semi/lynk/function/notice_board/model/dto/NoticeDTO.java
@@ -1,0 +1,4 @@
+package com.semi.lynk.function.notice_board.model.dto;
+
+public class NoticeDTO {
+}
