+++ src/main/java/com/semi/lynk/function/notice_board/service/NoticeService.java
@@ -1,0 +1,4 @@
+package com.semi.lynk.function.notice_board.service;
+
+public class NoticeService {
+}
