+++ src/main/java/com/semi/lynk/function/sales_management/controller/SalesController.java
@@ -1,0 +1,4 @@
+package com.semi.lynk.function.sales_management.controller;
+
+public class SalesController {
+}
