+++ src/main/java/com/semi/lynk/function/sales_management/model/dao/SalesMapper.java
@@ -1,0 +1,4 @@
+package com.semi.lynk.function.sales_management.model.dao;
+
+public interface SalesMapper {
+}
