+++ src/main/java/com/semi/lynk/function/sales_management/model/dto/SalesDTO.java
@@ -1,0 +1,4 @@
+package com.semi.lynk.function.sales_management.model.dto;
+
+public class SalesDTO {
+}
