+++ src/main/java/com/semi/lynk/function/sales_management/service/SalesService.java
@@ -1,0 +1,4 @@
+package com.semi.lynk.function.sales_management.service;
+
+public class SalesService {
+}
