--- src/main/java/com/semi/lynk/model/dto/human/DepartmentDTO.java
@@ -1,15 +1,0 @@
-package com.semi.lynk.model.dto.human;
-
-import lombok.*;
-
-@NoArgsConstructor
-@AllArgsConstructor
-@Setter
-@Getter
-@ToString
-public class DepartmentDTO {
-
-    private int depNo;      // 부서번호
-    private String depName; // 부서이름
-
-}
