--- src/main/java/com/semi/lynk/model/dto/human/HumanDTO.java
@@ -1,28 +1,0 @@
-package com.semi.lynk.model.dto.human;
-
-import lombok.*;
-
-@NoArgsConstructor
-@AllArgsConstructor
-@Getter
-@Setter
-@ToString
-public class HumanDTO {
-
-    private double salary;            // 기본급
-    private double incentive;           // 인센티브
-    private float vacationDayOfAnnual; // 연차 발생 일수
-    private float yearOfUse;            // 사용 연차
-    private String position;            // 직책
-    private String phoneNumber;         // 휴대 전화
-    private String address;             // 주소
-    private String residentNumber;       // 주민 번호
-    private String MaritalStatus;       // 결혼 여부
-    private String dayOfEmployment;     // 입사일자
-    private String dayOfResignation;    // 퇴사일자 (null 허용)
-    private String nationality;         // 국적
-    private String employmentStatus;    // 고용 구분
-    private String salaryBank;          // 급여 은행
-    private String salaryAccount;       // 급여 계좌
-    private String text;                // 비고 (null 허용)
-}
