+++ src/main/resources/templates/common/main.html
@@ -1,0 +1,16 @@
+<!doctype html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport"
+          content="width=device-width, user-scalable=no, initial-scale=1.0, maximum-scale=1.0, minimum-scale=1.0">
+    <meta http-equiv="X-UA-Compatible" content="ie=edge">
+    <title>Document</title>
+</head>
+<body> <!-- 최초 메인 페이지 -->
+    <h1>최초 화면</h1>
+    <button onclick="location.href='/employee/list'">사원 전체 조회</button>
+    <button onclick="location.href='/employee/join'">사원&부서 전체 조회</button>
+    <button onclick="location.href='/employee/regist'">인사 등록</button>
+</body>
+</html>
