+++ src/main/resources/templates/function/human/registPage.html
